Initial commit of deploy repository
This splits out the docker-compose file into a separate repository
which will be the only code pulled down from the web server. The
docker-compose contains the instructions on how to launch the
different services of the webapp. The services will be pulled down
from built container images that are located in our registry.

For now we will be using a simple .env file on the webserver to
specify the environment variables needed for all of the services.