Update docker-compose to pull service images
This commit updates the docker-compose configuration to pull the
backend and frontend container images from the container registry
instead of building them. We make this change now that we split the
web app into 4 independent repos instead of having 1 with all of
the source code.

The backend and frontend containers will be built on the dev server
and pushed up to the registry. They will be available for the
webserver to pull them down.