Update docker-compose to pull images instead of build
This updates the docker-compose file to pull images from the
registry rather than assume a build file will be present. This is
now necessary since each service is split off into their own
repositories, and we don't want to pull the source code for each
one to the webserver. So now, an image will be built for each
service and stored in the registry. The webserver only needs this
repository and can pull the necessary docker images down.