Add log bind mount to nginx service
This updates the docker-compose to add a bind mount to the nginx
service so that logs are preserved across containers. We follow the
same pattern as the backend service and store logs in the local
log/ directory.

We also change the port mapping from inside the container to be
8080 rather than 80 since we recently updated the nginx service to
run as a non-root user. This means the nginx image cannot bind to
80 since it's non-root, that this mapping takes care of that for
us.