Add Redis and Sidekiq, update pg17.4
This updates the docker compose file to add two new services: Redis
and Sidekiq. Sidekiq will use the backend rails image but run a
different command to start the sidekiq server.

We also update the postgres image to 17.4 which is the latest.

And finally, we make the deploy more secure by splitting the
networks into two: backend and frontend network. The backend
services will run on the backend network, and the frontend app will
run on the frontend network and doesn't need access to any backend
services like redis, postgres, sidekiq. Nginx has access to both
networks as that is the front door reverse proxy for all requests
to our app.