Update Rails secrets
Updates the environment variables to pass in the SECRET_KEY_BASE
directly to the Rails app rather than store it in the application
credentials encrypted file. For production, the secret key base was
the only secret stored in the application credentials. Everything
else is passed in as an environment variable. So lets just remove
this encrypted file for production all together.